Adds dirs to make project scallable
