+++ lib/screens/login.dart
@@ -1,0 +1,119 @@
+import 'package:flutter/material.dart';
+
+class LoginPage extends StatefulWidget {
+  LoginPage({Key key, this.title}) : super(key: key);
+  // This widget is the home page of your application. It is stateful, meaning
+  // that it has a State object (defined below) that contains fields that affect
+  // how it looks.
+  // This class is the configuration for the state. It holds the values (in this
+  // case the title) provided by the parent (in this case the App widget) and
+  // used by the build method of the State. Fields in a Widget subclass are
+  // always marked "final".
+  final String title;
+  @override
+  _LoginPageState createState() => _LoginPageState();
+}
+
+class _LoginPageState extends State<LoginPage> {
+
+   final emailController = TextEditingController();
+   final passwordController = TextEditingController();
+
+  @override
+  void dispose() {
+    // Clean up the controller when the widget is disposed.
+    emailController.dispose();
+    passwordController.dispose();
+    super.dispose();
+  }
+
+      @override
+      Widget build(BuildContext context) {
+
+        final emailField = TextField(
+          controller: TextEditingController(),
+          obscureText: false,
+          decoration: InputDecoration(
+              filled: true,
+              fillColor: Colors.white,
+              contentPadding: EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
+              hintText: "Email",
+              labelStyle: new TextStyle(color: Colors.white),
+              border:
+                  OutlineInputBorder(borderRadius: BorderRadius.circular(32.0))),
+        );
+
+        final passwordField = TextField(
+          controller: TextEditingController(),
+          obscureText: true,
+          decoration: InputDecoration(
+              filled: true,
+              fillColor: Colors.white,
+              contentPadding: EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
+              hintText: "Password",
+              border:
+                  OutlineInputBorder(borderRadius: BorderRadius.circular(32.0))),
+        );
+
+        final loginButon = Material(
+          elevation: 5.0,
+          borderRadius: BorderRadius.circular(30.0),
+          color: Colors.orange[200],
+          child: MaterialButton(
+            minWidth: MediaQuery.of(context).size.width,
+            padding: EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
+            onPressed: () {
+          return showDialog(
+            context: context,
+            builder: (context) {
+              return AlertDialog(
+                // Retrieve the text the that user has entered by using the
+                // TextEditingController.
+                content: Text(emailController.text),
+              );
+            },
+          );
+        },
+            child: Text("Login",
+                textAlign: TextAlign.center,
+          ),
+        )
+        );
+
+        return Scaffold(
+          body: Center(
+            child: Container(
+              color: Colors.cyan[900],
+              child: 
+              Padding(
+                padding: const EdgeInsets.all(36.0),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: <Widget>[
+                    SizedBox(
+                      height: 55.0,
+                      child: Image.asset(
+                        "assets/logo_enei.png",
+                        fit: BoxFit.contain,
+                      ),
+                    ),
+                    SizedBox(height: 20.0),
+                    emailField,
+                    SizedBox(height: 20.0),
+                    passwordField,
+                    SizedBox(
+                      height: 15.0,
+                    ),
+                    loginButon,
+                    SizedBox(
+                      height: 15.0,
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+        );
+      }
+    }
